+++ osidb/sync_manager.py
@@ -1,5 +1,6 @@
 import json
 from datetime import timedelta
+from typing import Optional, Type
 
 from celery.exceptions import Ignore
 from celery.utils.log import get_task_logger
@@ -64,18 +65,34 @@ class SyncManager(models.Model):
         raise NotImplementedError("Update synced links not implemented.")
 
     @classmethod
-    def schedule(cls, sync_id, *args, **kwargs):
+    def check_conflicting_sync_managers(
+        cls, sync_id, celery_task, related_managers: list[Type["SyncManager"]]
+    ):
+        """
+        Override this method to check for conflicting sync managers.
+        """
+        raise NotImplementedError("Conflicting sync managers check not implemented.")
+
+    @classmethod
+    def schedule(cls, sync_id, *args, schedule_options=None, **kwargs):
         """
         Schedule sync_task to Celery queue.
 
         :param sync_id: Unique ID for synchronized data object.
+        :param schedule_options: Dictionary of options to pass to apply_async
         """
+
+        if schedule_options is None:
+            schedule_options = {}
+
         cls.objects.get_or_create(sync_id=sync_id)
         cls.objects.filter(sync_id=sync_id).update(last_scheduled_dt=timezone.now())
 
         def schedule_task():
             try:
-                cls.sync_task.apply_async(args=[sync_id, *args], kwargs=kwargs)
+                cls.sync_task.apply_async(
+                    args=[sync_id, *args], kwargs=kwargs, **schedule_options
+                )
             except AttributeError:
                 raise NotImplementedError(
                     "Sync task not implemented or not implemented as Celery task."
@@ -87,13 +104,31 @@ class SyncManager(models.Model):
         transaction.on_commit(schedule_task)
 
     @classmethod
-    def started(cls, sync_id, celery_task):
+    def started(
+        cls,
+        sync_id,
+        celery_task,
+        related_managers: Optional[list[Type["SyncManager"]]] = None,
+    ):
         """
         This method has to be called at the beginning of the sync_task.
 
         :param sync_id: Unique ID for synchronized data object.
         :param celery_task: Associated Celery task.
+        :param related_managers: Optional related managers which should be checked for conflicts
         """
+
+        if related_managers is None:
+            related_managers = []
+
+        try:
+            cls.check_conflicting_sync_managers(sync_id, celery_task, related_managers)
+        except NotImplementedError:
+            logger.info(
+                f"{cls.__name__} {sync_id}: "
+                "Conflicting sync managers check not implemented"
+            )
+
         manager = cls.objects.get(sync_id=sync_id)
 
         # Check if task should really run, maybe it was scheduled more times
@@ -648,9 +683,12 @@ class JiraTaskDownloadManager(SyncManager):
         from collectors.jiraffe.convertors import JiraTaskConvertor
         from collectors.jiraffe.core import JiraQuerier
 
-        JiraTaskDownloadManager.started(task_id, self)
-
         set_user_acls(settings.ALL_GROUPS)
+        JiraTaskDownloadManager.started(
+            task_id,
+            self,
+            related_managers=[JiraTaskSyncManager, JiraTaskTransitionManager],
+        )
 
         try:
             task_data = JiraQuerier().get_issue(task_id, expand="changelog")
@@ -661,6 +699,42 @@ class JiraTaskDownloadManager(SyncManager):
             JiraTaskDownloadManager.failed(task_id, e)
         else:
             JiraTaskDownloadManager.finished(task_id)
+
+    @classmethod
+    def check_conflicting_sync_managers(
+        cls, sync_id, celery_task, related_managers: list[Type[SyncManager]]
+    ):
+        from osidb.models import Flaw
+
+        set_user_acls(settings.ALL_GROUPS)
+        countdown = 60  # 1 minute
+        existing_flaw = Flaw.objects.filter(task_key=sync_id).first()
+
+        if existing_flaw:
+            for sync_manager_type in related_managers:
+                conflicting_sync_manager = sync_manager_type.objects.filter(
+                    sync_id=existing_flaw.uuid
+                ).first()
+
+                if conflicting_sync_manager:
+                    conflicting_pending_sync = (
+                        conflicting_sync_manager.last_scheduled_dt is not None
+                        and (
+                            conflicting_sync_manager.last_finished_dt is None
+                            or conflicting_sync_manager.last_finished_dt
+                            < conflicting_sync_manager.last_scheduled_dt
+                        )
+                    )
+
+                    if conflicting_pending_sync:
+                        logger.info(
+                            f"{cls.__name__} {sync_id}: Conflicting sync managers found "
+                            f"({conflicting_sync_manager.__class__.__name__}), postponed for "
+                            f"{countdown} seconds."
+                        )
+                        cls.schedule(sync_id, schedule_options={"countdown": countdown})
+                        manager = cls.objects.get(sync_id=sync_id)
+                        manager.revoke_sync_task(celery_task)
 
     def update_synced_links(self):
         from osidb.models import Flaw