+++ docs/CHANGELOG.md
@@ -8,6 +8,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Add feedback form to tracker description (OSIDB-4095)
 
+### Fixed
+- Postpone Jira task download when there are sync or transition managers pending (OSIDB-4136)
+
 ## [4.10.0] - 2025-03-25
 ### Fixed
 - Make task_key read-only (OSIDB-4080)